+++ crates/ryft-pjrt/src/lib.rs
@@ -200,6 +200,7 @@ pub(crate) mod ffi {
     pub const PJRT_Extension_Type_AbiVersion: PJRT_Extension_Type = 20;
     pub const PJRT_Extension_Type_Collectives: PJRT_Extension_Type = 21;
     pub const PJRT_Extension_Type_MultiSlice: PJRT_Extension_Type = 22;
+    pub const PJRT_Extension_Type_HostMemoryAllocator: PJRT_Extension_Type = 23;
 
     /// PJRT extension base type. The `extension_type` field must be used to identify the type of the extension
     /// and reinterpret its instance accordingly.
@@ -571,7 +572,7 @@ mod tests {
 
         let plugin = test_cpu_plugin();
         let api = plugin.api();
-        assert_eq!(plugin.attribute("stablehlo_current_version"), Ok(Value::i64_list([1, 15, 0])));
+        assert_eq!(plugin.attribute("stablehlo_current_version"), Ok(Value::i64_list([1, 16, 0])));
         assert_eq!(plugin.attribute("stablehlo_minimum_version"), Ok(Value::i64_list([0, 9, 0])));
         assert_eq!(plugin.attribute("xla_version"), Ok(Value::i64(2)));
         assert_eq!(plugin.attribute("xla_version"), api.attribute("xla_version"));
@@ -579,7 +580,7 @@ mod tests {
             plugin.attribute("__missing__"),
             Err(Error::NotFound { message, .. }) if message.contains("__missing__")));
         let attributes = plugin.attributes().unwrap();
-        assert_eq!(attributes.get("stablehlo_current_version"), Some(&Value::i64_list([1, 15, 0])));
+        assert_eq!(attributes.get("stablehlo_current_version"), Some(&Value::i64_list([1, 16, 0])));
         assert_eq!(attributes.get("stablehlo_minimum_version"), Some(&Value::i64_list([0, 9, 0])));
         assert_eq!(attributes.get("xla_version"), Some(&Value::i64(2)));
         assert_eq!(attributes.get("__missing__"), None);